backend/requirements-api.txt
fastapi==0.115.12
uvicorn==0.34.0
pandas==2.2.3
SQLAlchemy==2.0.40
pydantic==2.13.3
pytesseract
pdf2image
python-docx
pillow
python-multipart
